+++ tests/Guzzle/Tests/Http/ClientTest.php
@@ -225,6 +225,7 @@ class ClientTest extends \Guzzle\Tests\GuzzleTestCase
 
     public function testClientAddsParamsToRequests()
     {
+        Version::$emitWarnings = false;
         $client = new Client('http://www.example.com', array(
             'api' => 'v1',
             'request.params' => array(
@@ -235,6 +236,7 @@ class ClientTest extends \Guzzle\Tests\GuzzleTestCase
         $request = $client->createRequest();
         $this->assertEquals('bar', $request->getParams()->get('foo'));
         $this->assertEquals('jar', $request->getParams()->get('baz'));
+        Version::$emitWarnings = true;
     }
 
     public function urlProvider()